lv_porting/fbdev: clear notify flag when flush ok
VELAPLATFO-7926

When using single buffer mode, FBIOPAN_DISPLAY will not be called, causing the pollready flag in fb to not be cleared in time. It will cause the poll to succeed multiple times in a vsync signal and render repeatedly. Call FBIO_CLEARNOTIFY at the end of lvgl flush to clear the writable flag to ensure that rendering will not be repeated.

Change-Id: I68ed096f43d83b16a55145f715d133ad6eb9f449
Signed-off-by: pengyiqiang <pengyiqiang@xiaomi.com>